+++ lib/rufo/file_finder.rb
@@ -1,36 +1,48 @@
 require "find"
+require "rake/file_list"
 
 class Rufo::FileFinder
   include Enumerable
 
   # Taken from https://github.com/ruby/rake/blob/f0a897e3fb557f64f5da59785b1a4464826f77b2/lib/rake/application.rb#L41
   RAKEFILES = [
-    "rakefile",
-    "Rakefile",
-    "rakefile.rb",
-    "Rakefile.rb",
+    "**/rakefile",
+    "**/Rakefile",
+    "**/rakefile.rb",
+    "**/Rakefile.rb",
   ]
 
   FILENAMES = [
-    "Gemfile",
+    "**/Gemfile",
     *RAKEFILES,
   ]
 
   EXTENSIONS = [
-    ".rb",
-    ".gemspec",
-    ".rake",
-    ".ru",
-    ".jbuilder",
-    ".erb",
+    "**/*.rb",
+    "**/*.gemspec",
+    "**/*.rake",
+    "**/*.ru",
+    "**/*.jbuilder",
+    "**/*.erb",
+  ]
+
+  DEFAULT_PATTERNS = [
+    *FILENAMES,
+    *EXTENSIONS,
   ]
 
   EXCLUDED_DIRS = [
     "vendor",
   ]
 
-  def initialize(files_or_dirs)
+  EXCLUDE_PATTERNS = [
+    "vendor/**/*",
+  ]
+
+  def initialize(files_or_dirs, includes: [], excludes: [])
     @files_or_dirs = files_or_dirs
+    @includes = includes
+    @excludes = excludes
   end
 
   def each
@@ -45,20 +57,15 @@ class Rufo::FileFinder
 
   private
 
-  attr_reader :files_or_dirs
+  attr_reader :files_or_dirs, :includes, :excludes
 
   def all_rb_files(file_or_dir)
-    files = []
-    Find.find(file_or_dir) do |path|
-      basename = File.basename(path)
-      if File.directory?(path)
-        Find.prune if EXCLUDED_DIRS.include?(basename)
-      else
-        if EXTENSIONS.include?(File.extname(basename)) || FILENAMES.include?(basename)
-          files << path
-        end
-      end
+    Dir.chdir(file_or_dir) do
+      fl = Rake::FileList.new(*DEFAULT_PATTERNS)
+      fl.exclude(*EXCLUDE_PATTERNS)
+      fl.exclude(*excludes)
+      fl.include(*includes)
+      fl.to_a
     end
-    files
   end
 end